requirements.txt
torch
numpy
opencv-python
matplotlib
nibabel
pillow
tiatooolbox
scikit-image
